intial

+++ src/main/java/com/robson/pride/api/skillcore/SkillCore.java
@@ -25,7 +25,8 @@ public class SkillCore {
         if (ent != null) {
             if (ent.getMainHandItem().getTag().getBoolean("hasweaponart")) {
                 weaponArtCore(ent, ent.getMainHandItem().getTag().getShort("weapon_art"));
-            } else defaultSkillCore(ent, ent.getMainHandItem());
+            }
+            else defaultSkillCore(ent, ent.getMainHandItem());
         }
     }
 
@@ -63,7 +64,7 @@ public class SkillCore {
     public static void loopParticleHit(Entity dmgent, Entity target, Particle particle, List<Entity> hitentities, float particleradius, Runnable function) {
         if (dmgent != null && target != null && particle != null && hitentities != null && function != null) {
             if (canHit(dmgent, target, hitentities)) {
-                if (MathUtils.getTotalDistance(particle.getPos().x - target.getX(), particle.getPos().y - target.getY(), particle.getPos().z - target.getZ()) <= particleradius + (target.getBbWidth() / 2)) {
+                if (MathUtils.getTotalDistance((float) (particle.getPos().x - target.getX()), (float) (particle.getPos().y - target.getY()), (float) (particle.getPos().z - target.getZ())) <= particleradius + (target.getBbWidth() / 2)) {
                     hitentities.add(target);
                     function.run();
                 } else
